Title: 합과 평균의 차, Time: 49ms, Memory: 16MB, Status: Passed - Codetree

+++ 250630/합과 평균의 차/sub-of-average-and-sum.py
@@ -1,4 +1,4 @@
 a, b, c = map(int, input().split())
 s = a+b+c
-avg = s/3
+avg = int(s/3)
 print(s, avg, s-avg, sep="\n")
